# apps/api-fastapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nha Tro Manager API",
    "description": "H\u1ec7 th\u1ed1ng RESTful API Qu\u1ea3n l\u00fd Nh\u00e0 tr\u1ecd theo ki\u1ebfn tr\u00fac T\u00e1ch bi\u1ec7t FastAPI & React SPA",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/contracts": {
      "get": {
        "tags": [
          "Contracts"
        ],
        "summary": "Get Contracts",
        "operationId": "get_contracts_api_v1_contracts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "buildingId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "L\u1ecdc theo T\u00f2a nh\u00e0",
              "title": "Buildingid"
            },
            "description": "L\u1ecdc theo T\u00f2a nh\u00e0"
          },
          {
            "name": "roomId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "L\u1ecdc theo Ph\u00f2ng tr\u1ecd",
              "title": "Roomid"
            },
            "description": "L\u1ecdc theo Ph\u00f2ng tr\u1ecd"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ContractStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "L\u1ecdc theo Tr\u1ea1ng th\u00e1i",
              "title": "Status"
            },
            "description": "L\u1ecdc theo Tr\u1ea1ng th\u00e1i"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Pagesize"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_PaginatedData_ContractResponseSchema__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Contracts"
        ],
        "summary": "Create Contract",
        "operationId": "create_contract_api_v1_contracts_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContractCreateSchema"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_ContractResponseSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/contracts/{contract_id}/tenants": {
      "get": {
        "tags": [
          "Contract Tenants"
        ],
        "summary": "Get Contract Tenants",
        "operationId": "get_contract_tenants_api_v1_contracts__contract_id__tenants_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Contract Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_List_ContractTenantResponseSchema__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Contract Tenants"
        ],
        "summary": "Add Tenant To Contract",
        "operationId": "add_tenant_to_contract_api_v1_contracts__contract_id__tenants_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Contract Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContractTenantCreateSchema"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_ContractTenantResponseSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contracts/{contract_id}/tenants/{tenant_id}": {
      "delete": {
        "tags": [
          "Contract Tenants"
        ],
        "summary": "Remove Tenant From Contract",
        "operationId": "remove_tenant_from_contract_api_v1_contracts__contract_id__tenants__tenant_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Contract Id"
            }
          },
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contracts/{contract_id}/change-primary": {
      "post": {
        "tags": [
          "Contract Tenants"
        ],
        "summary": "Change Primary Tenant",
        "operationId": "change_primary_tenant_api_v1_contracts__contract_id__change_primary_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Contract Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePrimaryTenantSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIResponse_ContractResponseSchema_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContractResponseSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "requestId": {
            "type": "string",
            "title": "Requestid"
          }
        },
        "type": "object",
        "title": "APIResponse[ContractResponseSchema]"
      },
      "APIResponse_ContractTenantResponseSchema_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContractTenantResponseSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "requestId": {
            "type": "string",
            "title": "Requestid"
          }
        },
        "type": "object",
        "title": "APIResponse[ContractTenantResponseSchema]"
      },
...
      "APIResponse_List_ContractTenantResponseSchema__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ContractTenantResponseSchema"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "requestId": {
            "type": "string",
            "title": "Requestid"
          }
        },
        "type": "object",
        "title": "APIResponse[List[ContractTenantResponseSchema]]"
      },
...
      "APIResponse_PaginatedData_ContractResponseSchema__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginatedData_ContractResponseSchema_"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "requestId": {
            "type": "string",
            "title": "Requestid"
          }
        },
        "type": "object",
        "title": "APIResponse[PaginatedData[ContractResponseSchema]]"
      },
...
      "APIResponse_dict_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "requestId": {
            "type": "string",
            "title": "Requestid"
          }
        },
        "type": "object",
        "title": "APIResponse[dict]"
      },
...
      "ChangePrimaryTenantSchema": {
        "properties": {
          "newPrimaryTenantId": {
            "type": "string",
            "title": "Newprimarytenantid",
            "description": "ID Kh\u00e1ch thu\u00ea \u0111\u01b0\u1ee3c ch\u1ecdn l\u00e0m \u0111\u1ea1i di\u1ec7n m\u1edbi (PRIMARY)"
          }
        },
        "type": "object",
        "required": [
          "newPrimaryTenantId"
        ],
        "title": "ChangePrimaryTenantSchema"
      },
...
      "ContractCreateSchema": {
        "properties": {
          "roomId": {
            "type": "string",
            "title": "Roomid",
            "description": "ID Ph\u00f2ng tr\u1ecd"
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "title": "Startdate",
            "description": "Ng\u00e0y b\u1eaft \u0111\u1ea7u h\u1ee3p \u0111\u1ed3ng"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "title": "Enddate",
            "description": "Ng\u00e0y k\u1ebft th\u00fac h\u1ee3p \u0111\u1ed3ng"
          },
          "monthlyPrice": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Monthlyprice",
            "description": "Gi\u00e1 thu\u00ea h\u00e0ng th\u00e1ng (> 0 VN\u0110)"
          },
          "depositAmount": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Depositamount",
            "description": "Ti\u1ec1n c\u1ecdc (>= 0 VN\u0110)",
            "default": "0"
          },
          "billingDay": {
            "type": "integer",
            "maximum": 28.0,
            "minimum": 1.0,
            "title": "Billingday",
            "description": "Ng\u00e0y ch\u1ed1t ti\u1ec1n ph\u00f2ng h\u00e0ng th\u00e1ng (1-28)",
            "default": 1
          },
          "paymentCycleMonths": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Paymentcyclemonths",
            "description": "K\u1ef3 h\u1ea1n thanh to\u00e1n (th\u00e1ng)",
            "default": 1
          },
          "terms": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terms"
          },
          "primaryTenantId": {
            "type": "string",
            "title": "Primarytenantid",
            "description": "ID Kh\u00e1ch thu\u00ea \u0111\u1ea1i di\u1ec7n (PRIMARY)"
          },
          "memberTenantIds": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Membertenantids",
            "description": "Danh s\u00e1ch Kh\u00e1ch thu\u00ea th\u00e0nh vi\u00ean (MEMBER)"
          }
        },
        "type": "object",
        "required": [
          "roomId",
          "startDate",
          "endDate",
          "monthlyPrice",
          "primaryTenantId"
        ],
        "title": "ContractCreateSchema"
      },
      "ContractResponseSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "contractCode": {
            "type": "string",
            "title": "Contractcode"
          },
          "roomId": {
            "type": "string",
            "title": "Roomid"
          },
          "roomNumber": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roomnumber"
          },
          "buildingName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buildingname"
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "title": "Startdate"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "title": "Enddate"
          },
          "actualEndDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actualenddate"
          },
          "monthlyPrice": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Monthlyprice"
          },
          "depositAmount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Depositamount"
          },
          "returnedDeposit": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Returneddeposit"
          },
          "depositDeductions": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Depositdeductions"
          },
          "billingDay": {
            "type": "integer",
            "title": "Billingday"
          },
          "paymentCycleMonths": {
            "type": "integer",
            "title": "Paymentcyclemonths"
          },
          "status": {
            "$ref": "#/components/schemas/ContractStatus"
          },
          "terms": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terms"
          },
          "cancellationReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancellationreason"
          },
          "terminationReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terminationreason"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "contractCode",
          "roomId",
          "startDate",
          "endDate",
          "monthlyPrice",
          "depositAmount",
          "returnedDeposit",
          "depositDeductions",
          "billingDay",
          "paymentCycleMonths",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "title": "ContractResponseSchema"
      },
      "ContractStatus": {
        "type": "string",
        "enum": [
          "DRAFT",
          "ACTIVE",
          "EXPIRING",
          "TERMINATED",
          "CANCELLED"
        ],
        "title": "ContractStatus"
      },
      "ContractTenantCreateSchema": {
        "properties": {
          "tenantId": {
            "type": "string",
            "title": "Tenantid",
            "description": "ID Kh\u00e1ch thu\u00ea"
          },
          "role": {
            "$ref": "#/components/schemas/ContractTenantRole",
            "description": "Vai tr\u00f2 (PRIMARY / MEMBER)",
            "default": "MEMBER"
          },
          "joinedAt": {
            "type": "string",
            "format": "date",
            "title": "Joinedat"
          }
        },
        "type": "object",
        "required": [
          "tenantId"
        ],
        "title": "ContractTenantCreateSchema"
      },
      "ContractTenantResponseSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "contractId": {
            "type": "string",
            "title": "Contractid"
          },
          "tenantId": {
            "type": "string",
            "title": "Tenantid"
          },
          "tenantName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenantname"
          },
          "tenantPhone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenantphone"
          },
          "role": {
            "$ref": "#/components/schemas/ContractTenantRole"
          },
          "joinedAt": {
            "type": "string",
            "format": "date",
            "title": "Joinedat"
          },
          "leftAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leftat"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "contractId",
          "tenantId",
          "role",
          "joinedAt",
          "createdAt"
        ],
        "title": "ContractTenantResponseSchema"
      },
      "ContractTenantRole": {
        "type": "string",
        "enum": [
          "PRIMARY",
          "MEMBER"
        ],
        "title": "ContractTenantRole"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedData_ContractResponseSchema_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ContractResponseSchema"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "pageSize": {
            "type": "integer",
            "title": "Pagesize"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "totalPages": {
            "type": "integer",
            "title": "Totalpages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "pageSize",
          "total",
          "totalPages"
        ],
        "title": "PaginatedData[ContractResponseSchema]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}